requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
djangorestframework-simplejwt
openai == 0.28
psycopg